--- PsychMonkey7/PsychMonkey7-api-impl/.gitignore
@@ -1,1 +1,0 @@
-/target/
