--- PsychMonkey7/PsychMonkey7-api-impl/.settings/org.eclipse.core.resources.prefs
@@ -1,5 +1,0 @@
-eclipse.preferences.version=1
-encoding//src/main/java=UTF-8
-encoding//src/main/resources=UTF-8
-encoding//src/test/java=UTF-8
-encoding/<project>=UTF-8
