--- PsychMonkey7/PsychMonkey7-api-impl/src/main/java/edu/asu/bmi/cad/PsychMonkey7/api/impl/DefaultHelloService.java
@@ -1,36 +1,0 @@
-package edu.asu.bmi.cad.PsychMonkey7.api.impl;
-
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-import edu.asu.bmi.cad.PsychMonkey7.api.HelloService;
-import org.springframework.stereotype.Service;
-
-/**
- * @author Olivier Lamy
- */
-@Service( "helloService#default" )
-public class DefaultHelloService
-    implements HelloService
-{
-    public String sayHello( String who )
-    {
-        return "Hello " + who;
-    }
-}
