--- PsychMonkey7/PsychMonkey7-api/.gitignore
@@ -1,1 +1,0 @@
-/target/
