--- PsychMonkey7/PsychMonkey7-api/.settings/org.eclipse.core.resources.prefs
@@ -1,3 +1,0 @@
-eclipse.preferences.version=1
-encoding//src/main/java=UTF-8
-encoding/<project>=UTF-8
