--- PsychMonkey7/PsychMonkey7-webapp/.settings/org.eclipse.core.resources.prefs
@@ -1,4 +1,0 @@
-eclipse.preferences.version=1
-encoding//src/main/resources=UTF-8
-encoding//src/test/java=UTF-8
-encoding/<project>=UTF-8
